+++ app/src/pages/Dashboard/EventsOverviewPage/fragments/EventsOverview.test.tsx
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: EUPL-1.2
 import React from 'react';
 
-import { createStore, eventMockedData, render, screen } from '../../../../utils/testUtils';
+import { configureStore, eventMockedData, render, screen } from '../../../../utils/testUtils';
 import { MeetingsProp } from '../EventsOverviewPage';
 import EventsOverview from './EventsOverview';
 
@@ -16,13 +16,13 @@ const mockEvents = [
 
 describe('EventsOverview', () => {
   test('page will not crash', async () => {
-    const { store } = createStore();
+    const { store } = configureStore();
     await render(<EventsOverview entries={mockEvents as MeetingsProp[]} expandAccordion={''} />, store);
     expect(screen.getAllByTestId('EventAccordion')).toHaveLength(1);
   });
 
   test('Accordion Title should be in the document', async () => {
-    const { store } = createStore();
+    const { store } = configureStore();
     await render(<EventsOverview entries={mockEvents as MeetingsProp[]} expandAccordion={'false'} />, store);
     expect(screen.getByText('Time Independent Meetings')).toBeInTheDocument();
   });